bitboxbase: Add a demo firmware for the HSM blinking LEDs.
Create a new folder "test/demo" which contains demo code.
This is going to be similar to the device-test folder, but
doesn't build firmwares with semihosting enabled (so firmwares
will startup immediately).

Move some of the CMake test code into a "common" CMakeLists
that builds the various libbitbox* libraries to be used
by all test code.