workflows: Allow for workflows to be run from the main loop.
Introduce a new workflow model, based on the workflow_t class.
This is the base class for all async workflows, and is managed
by a workflow_stack module. At each loop iteration, the top workflow
will be run. Replace the main "idle" screens with workflows matching
this semantics to allow the main loop to always have a workflow to spin.