components/confirm.c: Allow confirm buttons to take a custom callback parameter.
The UI callbacks should take a void* parameter. This allows the caller
to store context for later, without having to rely on the internal
representation of the component classes. At the moment all the
components use the ->parent links in the components they create to retrieve
their own data from callbacks - but this doesn't allow to store context
if the callback is not invoked from another component.