usb_processing: Better arbitration of the USB stack.
Create an arbiter in usb_processing.c, that allows one of the underlying subsystems (hww, bootloader, u2f) to put a lock on the USB stack (equivalent to what is currently happening with workflow_async_busy_set).

When the USB stack is locked, messages for other subsystems are rejected. Messages for the subsystem owning the lock are pre-filtered by asking the subsystem itself whether they should be allowed to process or not.

Each subsystem can define a custom handler for message rejection (so e.g. U2F will respond with U2F_SW_CONDITIONS_NOT_SATISFIED, and the HWW stack will respond with OP_ERR_DEVICE_BUSY).