USB processing: Start decoupling USB receive and USB sends.
USB sends can and will be sent only through a straight path starting
when a USB packet is received. This is inappropriate as it forces
processing the whole USB packet at once (hence blocking workflows etc).

Start moving the responsibility for packet sending out of the main
USB rx function - although the coupling is still there for now.